attr: map builtin userdiff drivers to well-known extensions
We already provide sane hunk-header patterns for specific languages.

However, the user has to manually map common extensions to use them. It's
not that hard to do, but it's an extra step that the user might not even
know is an option. Let's be nice and do it automatically.

It could be a problem in the future if the builtin userdiff drivers
started growing more invasive options, like automatically claiming to be
non-binary (i.e., setting diff.cpp.binary = false by default), but right
now we do not do that, so it should be safe.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>