path.c: increase array size of cwd variable to PATH_MAX
In make_absolute_path(), even though we use bufs[][PATH_MAX + 1] to keep
track of the partially resolved path in question, a separate buffer of a
smaller size cwd[1024] is given to getcwd() call. Cloning into a very deep
directory fails because of this.

The purpose of calling getcwd() is to save current working dir, then
before the make_absolute_path() returns, go back to the saved dir. We
should use PATH_MAX instead of 1024 for the array size.

Signed-off-by: Hui Wang <Hui.Wang@windriver.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>